refactor: improve file structure of gql in apps/api (#52)
* refactor(api): modularize GraphQL schema and loader structure

- Split GraphQL schema into multiple files (enums, types, queries, mutations, subscriptions)
- Added schema loader utility to merge .graphql files dynamically
- Centralized schema creation logic in a dedicated module
- Updated directory structure for better maintainability

* refactor(resolvers): split query and mutation resolvers into separate files for better modularity

Moved each main query and mutation resolver into its own file under queries and mutations
Updated main resolver files to import and compose from the new modular files
Updated import paths for Node16+ compatibility

* chore(api): update GraphQL tools dependencies and adjust module resolution settings

* fix: add .js extension to module imports for consistency

* feat: add script to copy GraphQL schema files during build process

* feat: add script to copy GraphQL schema files during build process
