+++ apps/api/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "dev": "tsx watch src/index.ts",
     "start": "node dist/index.js",
-    "build": "npm run db:generate && tsc",
+    "build": "npm run db:generate && tsc && node copy-graphql-schema.cjs",
     "lint": "eslint . --ext .ts,.tsx",
     "typecheck": "npm run db:generate && tsc -p tsconfig.json --noEmit",
     "format": "prettier . --write",
@@ -21,6 +21,8 @@
   "description": "",
   "dependencies": {
     "@apollo/server": "^4.12.2",
+    "@graphql-tools/load-files": "^7.0.1",
+    "@graphql-tools/merge": "^9.1.1",
     "@graphql-tools/schema": "^10.0.25",
     "@prisma/client": "^6.13.0",
     "body-parser": "^2.2.0",
