+++ apps/api/src/graphql/schema/subscriptions.graphql
@@ -1,0 +1,3 @@
+type Subscription {
+  newJob(minFit: Float = 10): Job!
+}
