+++ package.json
@@ -51,6 +51,8 @@
   },
   "dependencies": {
     "@dnd-kit/sortable": "^10.0.0",
+    "@graphql-tools/load-files": "^7.0.1",
+    "@graphql-tools/merge": "^9.1.1",
     "lodash": "^4.17.21"
   },
   "lint-staged": {
